Ensure failures in test prerequisites are not swallowed when TEST=all is passed to buildr.
A long standing bug existed such that if tests failed to compile, and the
option Buildr.options.test is set to :all then the compilation error would
not result in a failed build. As many people set the Buildr.options.test to
:all in their continuous delivery/integration tools, this has caused some
problems. This has been fixed by ensuring all the problems associated with
the resolving the prerequisites are not caught when Buildr.options.test is
set to :all.